requirements.txt
asgiref==3.7.2
Django==4.2.9
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
python-dotenv==1.0.0
pytz==2023.3.post1
psycopg2==2.8.6
sqlparse==0.4.4
tzdata==2023.4
whitenoise==6.6.0
